move static to example
