+++ doc/example/go.mod
@@ -1,0 +1,5 @@
+module github.com/utrack/clay/doc/example
+
+replace github.com/utrack/clay v0.0.0 => ../../
+
+require github.com/utrack/clay v0.0.0
