+++ doc/example/static/statik/statik.go
