+++ doc/example/static/swagger-ui-2.2.8/.jshintignore
